Added latests runs of the notebooks into the code folder
Added STL API for more accutate decomposition model.
Changed the threshold detection in the STL notebook from a constant to 3 std-devs of each time series.
Wrote  a proper set of tests comparing the performance of the old and new api in terms of percent correctness of number of anomalies. Also calculated F1, Precision and Recall for both approaches for A2, A3 and A4. A1 was giving me some trouble regarding the new API not detecting the frequency of the time series properly.